Refactored Sleep to use TimeSpan
git-svn-id: http://msbuildtasks.tigris.org/svn/msbuildtasks/trunk@56 299a232d-b705-0410-b782-f21f2f1e606a

+++ Source/MSBuild.Community.Tasks/Sleep.cs
@@ -101,7 +101,7 @@ namespace MSBuild.Community.Tasks
         private int GetSleepTime()
         {
             TimeSpan sleepTime = new TimeSpan(0, 0, _minutes, _seconds, _milliseconds);
-            return sleepTime.TotalMilliseconds;
+            return (int)sleepTime.TotalMilliseconds;
         }
     }
 }
